Refactor ecommerce items logic in separate class
